move cue to the stage component
